image-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
minio==7.2.9
python-dotenv==1.0.1
